+++ src/PYB11/Integrator/Integrator_PYB11.py
@@ -9,7 +9,7 @@ from SpheralCommon import *
 from spheralDimensions import *
 dims = spheralDimensions()
 
-HAVE_SOLVERS = "@ENABLE_SUNDIALS@"
+HAVE_SOLVERS = "@ENABLE_SOLVERS@"
 
 #-------------------------------------------------------------------------------
 # Includes