+++ src/PYB11/Solvers/EigenOptions.py
@@ -1,0 +1,9 @@
+from PYB11Generator import *
+
+@PYB11holder("std::shared_ptr")
+class EigenOptions:
+    def pyinit(self):
+        "Holds the options for Eigen solvers"
+
+    qr = PYB11readwrite()
+    reuseSolver = PYB11readwrite()